중복장비 기능추가, 인터페이스 분리
- 중복장비 기능추가, 인터페이스 분리
