+++ Battle.cs
@@ -141,97 +141,3 @@ public class BattleManager
         }
     }
 }
-
-public interface IAttack
-{
-    (int damage, bool isCritical, bool isMiss) Attack(Monster monster);
-}
-
-public class PlayerAttack : IAttack
-{
-    private readonly IDamageCalculator calculator;
-    public PlayerAttack(IDamageCalculator calculator)
-    {
-        this.calculator = calculator;
-    }
-    public (int damage, bool isCritical, bool isMiss) Attack(Monster monster)
-    {
-        var (damage, isCritical, isMiss) = calculator.Calculate(GameManager.Instance.player.TotalAttack, monster.def);
-        monster.health -= damage;
-
-        return (damage, isCritical, isMiss);
-    }
-}
-
-public class MonsterAttack : IAttack
-{
-    private readonly IDamageCalculator calculator;
-
-    public MonsterAttack(IDamageCalculator calculator)
-    {
-        this.calculator = calculator;
-    }
-
-    public (int damage, bool isCritical, bool isMiss) Attack(Monster monster)
-    {
-        var (damage, isCritical, isMiss) = calculator.Calculate(monster.attack, GameManager.Instance.player.TotalDefense);
-        GameManager.Instance.player.CurrentHP -= damage;
-        if (GameManager.Instance.player.CurrentHP < 0) GameManager.Instance.player.CurrentHP = 0;
-
-        return (damage, false, isMiss);
-    }
-}
-
-public interface IDamageCalculator
-{
-    (int damage, bool isCritical, bool isMiss) Calculate(int atk, int def);
-}
-public class PlayerBasicAttack : IDamageCalculator
-{
-    private Random random = new Random();
-    private double criticalChance = 0.15;
-    private double criticalHit = 1.6;
-    private double evasionChance = 0.5;
-
-    public (int damage, bool isCritical, bool isMiss) Calculate(int atk, int def)
-    {
-        bool isMiss = random.NextDouble() < evasionChance;
-        if (isMiss) return (0, false, true);
-
-        int error = (int)Math.Ceiling(atk * 0.1);
-        int min = atk - error;
-        int max = atk + error;
-        if (min < 0) min = 0;
-
-        int damage = random.Next(min, max + 1);
-        bool isCritical = random.NextDouble() < criticalChance;
-        if (isCritical) damage = (int)Math.Ceiling(damage * criticalHit);
-
-        int finalDamage = damage - def;
-        if (finalDamage < 0) finalDamage = 0;
-
-        return (finalDamage, isCritical, false);
-    }
-}
-public class MonsterBasicAttack : IDamageCalculator
-{
-    private Random random = new Random();
-    private double evasionChance = 0.5;
-
-    public (int damage, bool isCritical, bool isMiss) Calculate(int atk, int def)
-    {
-        bool isMiss = random.NextDouble() < evasionChance;
-        if (isMiss) return (0, false, true);
-
-        int error = (int)Math.Ceiling(atk * 0.1);
-        int min = atk - error;
-        int max = atk + error;
-        if (min < 0) min = 0;
-
-        int damage = random.Next(min, max + 1);
-        int finalDamage = damage - def;
-        if (finalDamage < 0) finalDamage = 0;
-
-        return (finalDamage, false, false);
-    }
-}
